% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banana.R
\name{banana}
\alias{banana}
\title{Banana Simulator}
\usage{
banana(x, variant = c("B", "A"))
}
\arguments{
\item{x}{Vector of input values, length d.}

\item{variant}{Type of multivariate extension.}
}
\value{
A scalar value, non-negative and almost certainly positive.
}
\description{
A general purpose scalar simulator based on extending Rosenbrock's banana function.  The two variants are "A" and "B", see Kok and Sandrock (2009).  "A" has a single stationary point at its global minimum of x = 1.  "B" has many stationary points, especially as the number of inputs becomes larger than about ten.  I have set the default variant to "B" as it is more interesting.

The banana function evaluates on the input space R^d, but in practice it suffices to consider [0, 1]^d.

See Examples for how the banana function can be turned into a toy simulator with both continuous and factor inputs.
}
\examples{
#### Banana simulator with mixed continuous and discrete inputs

## set up a template for the inputs

d <- 8L
temp <- lapply(1:d, function(j) {
  if (j == 3) {
    factor("dog", c("dog", "cat"))
  } else if (j == 6) {
    factor("apple", c("apple", "orange", "pear"))
  } else {
    runif(1L)
  }
})
temp <- as.data.frame(temp, col.names = LETTERS[1L:d])

## create the Banana simulator from this template

my_banana <- with(list(temp = temp), function(X) {
  runs <- check_runs(X, template = temp)
  X <- runs$X
  fac <- runs$fac
  for (j in fac) {
    x <- X[[j]]
    lev <- levels(x)
   X[[j]] <- as.numeric(x) / (1 + length(lev))
  }
  apply(as.matrix(X), 1L, banana)
})

## function to generate a random design matrix

rbana <- with(list(temp = temp), function(n) {
  X <- lapply(temp, function(x) {
    if (is.factor(x)) {
      lev <- levels(x)
      x <- lev[sample.int(length(lev), size = n, replace = TRUE)]
      factor(x, lev)
    } else {
      runif(n)
    }
  })
as.data.frame(X, col.names = LETTERS[1L:d])
})

## example

X <- rbana(21)
y <- my_banana(X)
runs <- check_runs(X, y, temp) # all OK

}
\references{
S. Kok and C. Sandrock, 2009, Locating and Characterizing the Stationary Points of the Extended Rosenbrock Function, \emph{Evolutionary Computing}, 17(3), 437-453.
}
