% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bopt.R
\name{bopt_qemu}
\alias{bopt_qemu}
\title{Bayesian Optimization}
\usage{
bopt_qemu(emu, Z, nz = 1000L, maximum = FALSE, npts = 31L)
}
\arguments{
\item{emu}{Object of class \code{qemu}, representing a function to be optimized.}

\item{Z}{Inputs over which to compute the expected improvement.}

\item{nz}{If \code{Z} is missing, generate \code{nz} points uniformly in the bounding box of the training dataset.}

\item{maximum}{Logical, maximize?  Default is to \emph{minimize}.}

\item{npts}{Number of points for the midpoint integration rule.}
}
\value{
A list with \code{Z} and \code{Eimp}, only for those rows where the expected improvement is positive.
}
\description{
Use an emulator of a function to score points in the input space according to their expected improvement (bigger is better).
}
\details{
The argument \code{nz} is provided for 'quick and dirty' optimization, but it will not generate candidate points outside the bounding box of the training dataset -- often this is where the optimum will lie.  It is often better to provide \code{Z} directly.
}
\examples{
## minimize the Banana function

\dontrun{
d <- 5
n <- 101
X <- matrix(runif(n * d), ncol = d)
colnames(X) <- LETTERS[1:d]
y <- apply(X, 1, banana)
emu <- make_qemu(X, y, inlogs = TRUE, nthreads = 2L)

bopt <- bopt_qemu(emu)
show(bopt) # only those with Eimp > 0

## let's have a look at the predictions

pp <- predict(emu, Z = bopt$Z)
pp$Eimp <- bopt$Eimp
show(pp)

## sequential minimization

collapse <- function(x) {
  paste(x, collapse = ", ")
}

nz <- 1E4 # more candidates
bopt <- bopt_qemu(emu, nz = nz)
niter <- 20
for (i in 1:niter) {
  x <- unlist(bopt$Z[1, ]) # Z is a dataframe
  X <- rbind(X, x)
  message("Sim run at ", collapse(signif(x, 3)))
  y <- c(y, banana(x))
  message("S(x) best so far = ", signif(min(y), 3))
  emu <- make_qemu(X, y, inlogs = TRUE)
  bopt <- bopt_qemu(emu)
  if (length(bopt$Eimp) == 0) {
    message("Early exit, no bopt run found")
    break
  }
}

}
}
\references{
W. I. Notz, 2015, Expected Improvement Designs. Ch19 in A. Dean et al, eds, Handbook of Design and Analysis of Experiments, Chapman & Hall, CRC.
}
