% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_runs.R
\name{check_runs}
\alias{check_runs}
\title{Check the Ensemble of Runs}
\usage{
check_runs(X, y = NULL, template = NULL)
}
\arguments{
\item{X}{Matrix or dataframe of input values.}

\item{y}{Vector or matrix of output values.}

\item{template}{Template for checking the inputs.}
}
\value{
If successful, a list with components \code{X} (dataframe) and \code{y} (numeric vector or matrix), and \code{factors} (character vector of names). Otherwise an error with an informative message.
}
\description{
Check the \code{X} and \code{y} (optional) of an ensemble of runs.
}
