% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRPS.R
\name{CRPS_qemu}
\alias{CRPS_qemu}
\title{Continuous Ranked Probability Score}
\usage{
CRPS_qemu(emu, X, y, npts = 51)
}
\arguments{
\item{emu}{Object of class \code{qemu}.}

\item{X, y}{Validation data, inputs and outputs.}

\item{npts}{Number of points for the midpoint integration rule.}
}
\value{
A vector of CRPS scores.
}
\description{
Compute the CRPS for a set of simulator runs.
}
\examples{
## see ?make_qemu
}
