% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOO.R
\name{LOO_qemu}
\alias{LOO_qemu}
\title{Leave-One-Out (LOO)}
\usage{
LOO_qemu(
  emu,
  subset = "all",
  type = c("full", "quick"),
  level = 0.95,
  plotit = TRUE,
  ...
)
}
\arguments{
\item{emu}{Object of class \code{QEMU}.}

\item{subset}{The rows to leave out, or \code{"all"} to do all rows (EXPENSIVE!).}

\item{type}{Full LOO (EXPENSIVE!) or quick-and-dirty?}

\item{level}{Coverage of the prediction interval.}

\item{plotit}{Logical, plot the result?}

\item{...}{Passed to \code{\link[=plot.LOO]{plot.LOO()}}.}
}
\value{
A dataframe with \code{actual}, \code{predict}, \code{lower}, and \code{upper} components, and a \code{X} attribute with the input values for each row.
}
\description{
LOO analysis for the emulator.  Since random forests are a bit forgetful, a similar result follows from just using the predict method, or \code{type = "quick"} in the arguments.
}
\examples{
## see ?make_qemu

}
