% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEFF.R
\name{meffs}
\alias{meffs}
\alias{MEFF_qemu}
\alias{MEFF_runs}
\title{Compute Main Effects}
\usage{
MEFF_qemu(object, inom = NA, level = 0.95, npts = 51, plotit = TRUE, ...)

MEFF_runs(X, y, level = 0.95, npts = 51, plotit = TRUE, ...)
}
\arguments{
\item{object}{Object of class \code{QEMU}.}

\item{inom}{Row-number of the nominal value in the training ensemble.}

\item{level}{Coverage of the pointwise prediction interval.}

\item{npts}{Number of points in each main effect.}

\item{plotit}{Logical, produce a Main Effects plot?}

\item{...}{arguments passed to \code{\link[=plot.MEFF]{plot.MEFF()}}, used if \code{plotit = TRUE}.}

\item{X, y}{Runs for \code{MEFF_runs()}.}
}
\value{
\code{MEFF_runs()} and \code{MEFF_qemu()} create an object of class \code{MEFF}, mainly for its side-effect of producing a Main Effects plot via the plot method.  See \code{\link[=plot.MEFF]{plot.MEFF()}}.
}
\description{
The main effect of the jth input is the simulator output when all but the jth input are at their nominal values, and the jth input is varied across its range, taken to be from the lowest to highest values in the training dataset.
\itemize{
\item \code{MEFF_qemu()} : compute the MEFFs from a \code{qemu}.
\item \code{MEFF_runs()} : compute the MEFFS directly from the runs.
}

The first is likely to be clearer.  \code{MEFF_runs()} just does marginal fit using \code{\link[stats:loess]{stats::loess()}} with its default arguments.
}
\examples{
## See ?make_qemu

}
