% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_LOO.R
\name{plot.LOO}
\alias{plot.LOO}
\title{Leave-One-Out (LOO) plot}
\usage{
\method{plot}{LOO}(
  x,
  y,
  along = NULL,
  ylim = NULL,
  log = "",
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  cex = 1,
  ...
)
}
\arguments{
\item{x}{Object of class \code{LOO}.}

\item{y}{Ignored.}

\item{along}{How to order the x-axis, the name of an input or its index.  Default uses the same ordering as \code{subset}.}

\item{ylim, log, xlab, ylab, main, cex}{Usual plotting arguments.}

\item{...}{For compatibility with \code{\link[=plot]{plot()}}.}
}
\value{
Called for its side-effect of producing a Leave One Out (LOO) plot, returns \code{x} invisibly.
}
\description{
Take a \code{LOO} object and plot it.  A \code{LOO} object is a dataframe with \code{actual}, \code{predicted}, \code{lower}, and \code{upper} components.
}
\examples{
## see ?make_qemu

}
