% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MEFF.R
\name{plot.MEFF}
\alias{plot.MEFF}
\title{Main Effects Plot}
\usage{
\method{plot}{MEFF}(
  x,
  y,
  runs,
  select = 1L:d,
  ylim = NULL,
  log = "",
  main = NULL,
  cex.main = 1.2,
  nc = NA,
  col = NA,
  alpha = 0.3,
  ...
)
}
\arguments{
\item{x}{An object of class \code{MEFF}.}

\item{y}{Ignored.}

\item{runs}{Optional list with \code{X} and \code{y} components.  Underplots the main effects with the runs.  Add \code{pch}, \code{cex}, and \code{col} components to \code{runs} to overwrite default values.}

\item{select}{Which main effects to plot.}

\item{ylim, log, main, cex.main}{Usual plot arguments.}

\item{nc}{How many columns to use in the layout.}

\item{col}{Colours for each component in \code{select}.}

\item{alpha}{'Fade' for the colour of the uncertainty sleeve, use \code{alpha = 0} to suppress the sleeve.}

\item{...}{For compatiblity with \code{\link[=plot]{plot()}}.}
}
\value{
Called for its side-effect of producing a Main Effects plot, returns \code{x} invisibly.
}
\description{
Take a \code{MEFF} object and plot it.  A \code{MEFF} object is a named list of dataframes, one for each input.  Each dataframe has components \code{x}, \code{predict}, \code{lower}, \code{upper}.

To tweak the plot, adjust the arguments in \code{\link[=par]{par()}}.
}
\examples{
## See ?make_qemu

}
\seealso{
\code{\link[=MEFF_runs]{MEFF_runs()}} and \code{\link[=MEFF_qemu]{MEFF_qemu()}} create \code{MEFF} objects.
}
