% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.qemu}
\alias{predict.qemu}
\title{Predictions for the 'qemu' Emulator}
\usage{
\method{predict}{qemu}(object, Z, type = c("interval", "moments", "ecdf"), level = 0.95, ...)
}
\arguments{
\item{object}{Object of class \code{qemu}.}

\item{Z}{New inputs to predict, must have the same column names and columns types as \code{X}.}

\item{type}{Type of prediction, see above.}

\item{level}{Coverage of the prediction interval when \code{type = "interval"}.}

\item{...}{For compatibilty with \code{\link[=predict]{predict()}}.}
}
\value{
A dataframe or a list, see above.
}
\description{
Predict the simulator output at any input values.  Three types of prediction are available:
\itemize{
\item \code{type = "interval"} : a dataframe with \code{lower}, \code{median}, and \code{upper} columns.
\item \code{type = "moments"} : a dataframe with \code{mean} and \code{sd} columns, plus additional columns \code{meanlog} and \code{sdlog} if \code{object} is made using \code{inlogs = TRUE}.
\item \code{type = "ecdf"} : a list of empirical distribution functions.
}
}
\examples{
## see ?make_qemu

}
\seealso{
\code{\link[=CRPS_qemu]{CRPS_qemu()}} which uses \code{type = "ecdf"} to compute diagnostic information using the Continuous Ranked Probability Score.
}
