% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qemu-package.R
\docType{package}
\name{qemu}
\alias{qemu}
\title{qemu package}
\description{
Train an emulator on a set of computer simulator runs, in order to predict output values at new input values.  Uses linear regressors, and a random forest for the residual process.  Can handle factor inputs.
\itemize{
\item \code{make_qemu()} makes an emulator from a training dataset.
\item \code{predict()} predicts the simulator at new input values.
\item \code{CRPS_qemu()} scores the simulator on a validation dataset.
\item \code{tune_qemu()} tunes the two parameters \code{mtry} and \code{nodesize}.
\item \code{LOO_qemu()} does a leave-one-out validation analysis.
\item \code{MEFF_qemu()} computes the main effects.
\item \code{bopt_qemu()} is for Bayesian minimization.
}

A 'qemu' is a cross between a kiwi and an emu.
}
\author{
Jonathan Rougier
}
