% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeve.R
\name{sleeve}
\alias{sleeve}
\title{Add a 'Sleeve' to a Plot}
\usage{
sleeve(x, lower, upper, col = "black", alpha = 0.3)
}
\arguments{
\item{x}{Abscissa values.}

\item{lower, upper}{pointwise lower and upper values corresponding to \code{x}.}

\item{col}{colour of the sleeve.}

\item{alpha}{opacity of the sleeve.}
}
\value{
Called for its side-effect of adding a sleeve to a plot.
}
\description{
A 'sleeve' is a polygon showing, for example, pointwise uncertainty.  It is used in \code{\link[=plot.MEFF]{plot.MEFF()}}.
}
\examples{
## see ?make_qemu, used in plot.MEFF()

}
\seealso{
\code{\link[scales:alpha]{scales::alpha()}} for handling opacity.
}
