% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{tune_qemu}
\alias{tune_qemu}
\alias{plot.tuned_qemu}
\title{Tune 'mtry' and 'nodesize'}
\usage{
tune_qemu(
  emu,
  prop_train = 0.9,
  nrep = 10L,
  mtry_vals = 1L:min(d, 8),
  nodesize_vals = 1L:8,
  plotit = TRUE,
  ...
)

\method{plot}{tuned_qemu}(x, y, cex = 0.9, ...)
}
\arguments{
\item{emu}{Object of class \code{qemu}.}

\item{prop_train}{Proportion of the runs used for training.}

\item{nrep}{Number of replications for each combination of \code{mtry} and \code{nodesize}.}

\item{mtry_vals}{Values for \code{mtry}.}

\item{nodesize_vals}{Values for \code{nodesize}.}

\item{plotit}{Logical, plot the result?}

\item{x}{Object of class \code{tune_qemu} for plot method.}

\item{y, ...}{For compatiility with \code{\link[=plot]{plot()}}.}

\item{cex}{Expansion factor for text values in the plot.}
}
\value{
A list with components:
\itemize{
\item \code{itrain} : indices of the training dataset.
\item \code{grid} : grid of values for \code{mtry} and \code{nodesize}.
\item \code{crps} : mean CRPS score over the hold-out sample.
\item \code{best} : best values of \code{mtry} and \code{nodesize}, as a list.
\item \code{tuned_qemu} : qemu with best parameters, refitted for all the runs.
}
}
\description{
Compute CRPS on a hold-out sample for a range of values for \code{mtry} and \code{nodesize}, and optionally plot these in a grid.
}
\examples{
## see ?make_qemu
}
\seealso{
\code{\link[=CRPS_qemu]{CRPS_qemu()}} for calculating the CRPS.
}
