% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wingweight.R
\name{wing_weight}
\alias{wing_weight}
\title{Wing Weight Function}
\usage{
wing_weight(..., TYPE = c("dataframe", "matrix"), ALL = FALSE)
}
\arguments{
\item{...}{Input values, see Inputs and Examples.}

\item{TYPE}{Type of object for the inputs in the function output.}

\item{ALL}{Logical, return all the inputs or just the ones in \code{...}.}
}
\value{
A list with \code{X} and \code{y} components.
}
\description{
The \emph{wing weight} function using the implementation from \url{https://www.sfu.ca/~ssurjano/Code/wingweightr.html}, adapted to change two of the inputs into factors (also know as 'categorical' or 'discrete' inputs).
}
\section{Inputs}{

The named arguments in \code{...} are combined into a dataframe, and then the columns in that dataframe are used to overwrite the nominal values for each input.  The inputs and their default values and ranges are:

\tabular{lll}{
Name \tab Description \tab Values (nominal, range) \cr
Sw \tab Wing area (ft^2) \tab 174, 150 to 200 \cr
Wfw \tab Weight of fuel in wing (lb) \tab 252, 220 to 300 \cr
A \tab Aspect ratio \tab Factor, 'nom' and 'high' \cr
LamCaps \tab Quarter-chord sweep (deg) \tab 0, -10 to 10 \cr
q \tab Dynamic pressure at cruise (lb/ft^2) \tab 34, 16 to 45 \cr
lam \tab Taper ratio \tab Factor, 'nom', 'low', 'high' \cr
tc \tab Aerofoil thickness to chord ratio \tab 0.12, 0.08 to 0.18 \cr
Nz \tab Ultimate load factor \tab 3.8, 2.5 to 6 \cr
Wdg \tab Final design gross weight (lb) \tab 2000, 1700 to 2500 \cr
Wp \tab Paint weight (lb/ft^2) \tab 0.05, 0.025 to 0.08 \cr
}
}

\examples{
## nominal values

ww <- wing_weight() # ALL = TRUE in this case
ww <- wing_weight(TYPE = "matrix") # factors become numeric

## can be called with a subset of inputs

Sw <- seq(from = 150, to = 200, length.out = 11)
ww <- wing_weight(Sw = Sw)

## or a dataframe

X <- expand.grid(Sw = Sw, A = c("nom", "high"))
ww <- wing_weight(X)

## some other combo

ww <- wing_weight(X, LamCaps = 5)

## get back all the input values

ww <- wing_weight(X, LamCaps = 5, ALL = TRUE)

}
